Refactor & Clean up Sync params config implementation
