+++ android/eir/src/main/java/org/smartregister/fhircore/eir/EirConfigService.kt
@@ -19,23 +19,11 @@ package org.smartregister.fhircore.eir
 import android.content.Context
 import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
-import org.hl7.fhir.r4.model.ResourceType
 import org.smartregister.fhircore.engine.configuration.app.AuthConfiguration
 import org.smartregister.fhircore.engine.configuration.app.ConfigService
-import org.smartregister.fhircore.engine.configuration.app.RegistrySyncParamConfigService
 
 class EirConfigService @Inject constructor(@ApplicationContext val context: Context) :
-  ConfigService, RegistrySyncParamConfigService {
-
-  override val resourceSyncParams: Map<ResourceType, Map<String, String>>
-    get() =
-      mapOf(
-        ResourceType.Patient to emptyMap(),
-        ResourceType.Immunization to emptyMap(),
-        ResourceType.Questionnaire to emptyMap(),
-        ResourceType.StructureMap to mapOf(),
-        ResourceType.RelatedPerson to mapOf()
-      )
+  ConfigService {
 
   override fun provideAuthConfiguration() =
     AuthConfiguration(
