+++ android/engine/src/main/java/org/smartregister/fhircore/engine/configuration/app/RegistrySyncParamConfigService.kt
@@ -17,14 +17,6 @@
 package org.smartregister.fhircore.engine.configuration.app
 
 import com.google.android.fhir.FhirEngine
-import org.hl7.fhir.r4.model.Parameters
-import org.hl7.fhir.r4.model.ResourceType
-import org.hl7.fhir.r4.model.SearchParameter
-import org.smartregister.fhircore.engine.configuration.AppConfigClassification
-import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
-import org.smartregister.fhircore.engine.configuration.FhirConfiguration
-import org.smartregister.fhircore.engine.data.remote.model.response.UserInfo
-import timber.log.Timber
 
 /**
  * An interface that provides the application configurations.
@@ -33,70 +25,6 @@ import timber.log.Timber
  */
 interface RegistrySyncParamConfigService {
 
-  val resourceSyncParams: Map<ResourceType, Map<String, String>>
-
-  fun loadRegistrySyncParams(
-    configurationRegistry: ConfigurationRegistry,
-    authenticatedUserInfo: UserInfo?
-  ): Map<ResourceType, Map<String, String>> {
-    val pairs = mutableListOf<Pair<ResourceType, Map<String, String>>>()
-
-    val syncConfig =
-      configurationRegistry.retrieveConfiguration<FhirConfiguration<Parameters>>(
-        AppConfigClassification.SYNC
-      )
-
-    val appConfig =
-      configurationRegistry.retrieveConfiguration<ApplicationConfiguration>(
-        AppConfigClassification.APPLICATION
-      )
-
-    // TODO Does not support nested parameters i.e. parameters.parameters...
-    // TODO: expressionValue supports for Organization and Publisher literals for now
-    syncConfig.resource.parameter.map { it.resource as SearchParameter }.forEach { sp ->
-      val paramName = sp.name!! // e.g. organization
-      val paramLiteral = "#$paramName" // e.g. #organization in expression for replacement
-      val paramExpression = sp.expression
-      val expressionValue =
-        when (paramName) {
-          ConfigurationRegistry.ORGANIZATION -> authenticatedUserInfo?.organization
-          ConfigurationRegistry.PUBLISHER -> authenticatedUserInfo?.questionnairePublisher
-          ConfigurationRegistry.ID -> paramExpression
-          ConfigurationRegistry.COUNT -> appConfig.count
-          else -> null
-        }?.let {
-          // replace the evaluated value into expression for complex expressions
-          // e.g. #organization -> 123
-          // e.g. patient.organization eq #organization -> patient.organization eq 123
-          paramExpression.replace(paramLiteral, it)
-        }
-
-      // for each entity in base create and add param map
-      // [Patient=[ name=Abc, organization=111 ], Encounter=[ type=MyType, location=MyHospital ],..]
-      sp.base.forEach { base ->
-        val resourceType = ResourceType.fromCode(base.code)
-        val pair = pairs.find { it.first == resourceType }
-        if (pair == null) {
-          pairs.add(
-            Pair(
-              resourceType,
-              expressionValue?.let { mapOf(sp.code to expressionValue) } ?: mapOf()
-            )
-          )
-        } else {
-          expressionValue?.let {
-            // add another parameter if there is a matching resource type
-            // e.g. [(Patient, {organization=105})] to [(Patient, {organization=105, _count=100})]
-            val updatedPair = pair.second.toMutableMap().apply { put(sp.code, expressionValue) }
-            val index = pairs.indexOfFirst { it.first == resourceType }
-            pairs.set(index, Pair(resourceType, updatedPair))
-          }
-        }
-      }
-    }
-
-    Timber.i("SYNC CONFIG $pairs")
-
-    return mapOf(*pairs.toTypedArray())
-  }
+  /** Provide [AuthConfiguration] for the Application */
+  fun provideAuthConfiguration(): AuthConfiguration
 }
