+++ android/engine/src/main/java/org/smartregister/fhircore/engine/sync/SyncHandlers.kt
@@ -18,15 +18,21 @@ package org.smartregister.fhircore.engine.sync
 
 import androidx.lifecycle.LifecycleCoroutineScope
 import com.google.android.fhir.FhirEngine
-import com.google.android.fhir.sync.DownloadWorkManager
 import com.google.android.fhir.sync.State
 import com.google.android.fhir.sync.SyncJob
+import com.google.android.fhir.sync.download.ResourceParamsBasedDownloadWorkManager
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
+import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
+import org.smartregister.fhircore.engine.configuration.app.ConfigService
+import org.smartregister.fhircore.engine.data.remote.model.response.UserInfo
 import org.smartregister.fhircore.engine.util.DefaultDispatcherProvider
 import org.smartregister.fhircore.engine.util.DispatcherProvider
+import org.smartregister.fhircore.engine.util.SharedPreferencesHelper
+import org.smartregister.fhircore.engine.util.USER_INFO_SHARED_PREFERENCE_KEY
+import org.smartregister.fhircore.engine.util.extension.decodeJson
 import timber.log.Timber
 
 /**
@@ -44,7 +50,9 @@ interface OnSyncListener {
  * invoking [broadcastSync] method
  */
 class SyncBroadcaster(
-  val downloadWorkManager: DownloadWorkManager,
+  val configurationRegistry: ConfigurationRegistry,
+  val sharedPreferencesHelper: SharedPreferencesHelper,
+  val configService: ConfigService,
   val syncJob: SyncJob,
   val fhirEngine: FhirEngine,
   val sharedSyncStatus: MutableSharedFlow<State> = MutableSharedFlow(),
@@ -55,7 +63,18 @@ class SyncBroadcaster(
       try {
         syncJob.run(
           fhirEngine = fhirEngine,
-          downloadManager = downloadWorkManager,
+          downloadManager =
+            ResourceParamsBasedDownloadWorkManager(
+              syncParams =
+                configService
+                  .loadRegistrySyncParams(
+                    configurationRegistry,
+                    sharedPreferencesHelper
+                      .read(USER_INFO_SHARED_PREFERENCE_KEY, null)
+                      ?.decodeJson<UserInfo>()
+                  )
+                  .toMap()
+            ),
           subscribeTo = sharedSyncStatus
         )
       } catch (exception: Exception) {
