+++ android/engine/src/main/java/org/smartregister/fhircore/engine/util/extension/ApplicationExtension.kt
@@ -27,7 +27,16 @@ import com.google.android.fhir.search.StringFilterModifier
 import com.google.android.fhir.search.count
 import com.google.android.fhir.search.getQuery
 import com.google.android.fhir.search.search
+import com.google.android.fhir.sync.FhirSyncWorker
+import com.google.android.fhir.sync.PeriodicSyncConfiguration
+import com.google.android.fhir.sync.RepeatInterval
+import com.google.android.fhir.sync.SyncJob
 import com.google.gson.Gson
+import java.util.concurrent.TimeUnit
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.emitAll
+import kotlinx.coroutines.launch
 import org.hl7.fhir.r4.model.Bundle
 import org.hl7.fhir.r4.model.Immunization
 import org.hl7.fhir.r4.model.Library
@@ -35,8 +44,10 @@ import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.RelatedPerson
 import org.hl7.fhir.r4.model.Resource
 import org.hl7.fhir.r4.model.ResourceType
+import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
 import org.smartregister.fhircore.engine.cql.FhirOperatorDecorator
 import org.smartregister.fhircore.engine.data.domain.util.PaginationUtil
+import org.smartregister.fhircore.engine.sync.SyncBroadcaster
 import org.smartregister.fhircore.engine.util.SharedPreferencesHelper
 import timber.log.Timber
 
@@ -135,3 +146,22 @@ suspend fun FhirEngine.loadCqlLibraryBundle(
   } catch (exception: Exception) {
     Timber.e(exception)
   }
+
+/**
+ * Schedule periodic sync periodically as defined in the [configurationRegistry] application config
+ * interval. The [syncBroadcaster] will broadcast the sync status to its listeners
+ */
+fun SyncJob.schedulePeriodicSync(
+  configurationRegistry: ConfigurationRegistry, // TODO Obtain sync interval from app config
+  syncBroadcaster: SyncBroadcaster,
+  syncInterval: Long = 30
+) {
+  CoroutineScope(Dispatchers.Main).launch {
+    syncBroadcaster.sharedSyncStatus.emitAll(this@schedulePeriodicSync.stateFlow())
+  }
+  this.poll(
+    periodicSyncConfiguration =
+      PeriodicSyncConfiguration(repeat = RepeatInterval(syncInterval, TimeUnit.MINUTES)),
+    clazz = FhirSyncWorker::class.java
+  )
+}
