+++ android/quest/src/main/java/org/smartregister/fhircore/quest/di/config/ConfigServiceModule.kt
@@ -21,17 +21,10 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import org.smartregister.fhircore.engine.configuration.app.ConfigService
-import org.smartregister.fhircore.engine.configuration.app.RegistrySyncParamConfigService
 import org.smartregister.fhircore.quest.QuestConfigService
-import org.smartregister.fhircore.quest.QuestRegistrySyncParamConfigService
 
 @InstallIn(SingletonComponent::class)
 @Module
 abstract class ConfigServiceModule {
   @Binds abstract fun provideConfigService(questConfigService: QuestConfigService): ConfigService
-
-  @Binds
-  abstract fun provideRegistrySyncParamConfigService(
-    questRegistryConfigService: QuestRegistrySyncParamConfigService
-  ): RegistrySyncParamConfigService
 }
